examples/3-memory-and-context-window/pyproject.toml
[project]
name = "1-strctured-output"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.105.0",
    "python-dotenv>=1.1.1",
    "pydantic>=2.0.0",
    "faiss-cpu>=1.7.0",
    "sentence-transformers>=2.0.0",
]
